+++ apps/server/src/config.ts
@@ -5,12 +5,9 @@ const config = {
   rabbitMq: {
     url: process.env.RABBITMQ_URL ?? "amqp://localhost"
   },
-  serverHttp: {
+  serverApi: {
     baseUrl: "",
     port: parseInt(process.env.HTTP_PORT as string) ?? 3004
-  },
-  serverApollo: {
-    port: parseInt(process.env.APOLLO_PORT as string) ?? 4005
   },
   auth: {
     code: {
@@ -28,7 +25,7 @@ const config = {
   }
 };
 
-config.serverHttp.baseUrl =
-  process.env.HTTP_BASE_URL ?? `http://localhost:${config.serverHttp.port}`;
+config.serverApi.baseUrl =
+  process.env.HTTP_BASE_URL ?? `http://localhost:${config.serverApi.port}`;
 
 export default config;
